refactor: consolidate shape imports via math barrel export
- Add rectangle and triangle exports to math/index.ts
- Replace deep relative imports (../../math/shape/xxx) with barrel imports (../../math)
- Reorganize type imports in shape-component.ts to more specific paths
- Fix minor formatting (case indentation, EOF newline)
